+++ internal/web/webgui.go
@@ -35,6 +35,7 @@ func Gui(confPath, yamlPath, nodePath string) {
 	http.HandleFunc("/", indexHandler)                  // index.go
 	http.HandleFunc("/config/", configHandler)          // config.go
 	http.HandleFunc("/config_save/", saveConfigHandler) // config.go
+	http.HandleFunc("/file/", fileHandler)              // file.go
 	http.HandleFunc("/host/", hostHandler)              // host.go
 	http.HandleFunc("/panels/", panelsHandler)          // panels.go
 	http.HandleFunc("/panel_edit/", panelEditHandler)   // panel-edit.go